intermediate folder rename
