+++ src/_AmbientContext/AmbientContext.nuspec
@@ -1,0 +1,18 @@
+<?xml version="1.0"?>
+<package >
+  <metadata>
+    <id>AmbientContext</id>
+    <version>$version$</version>
+    <title>$title$</title>
+    <authors>$author$</authors>
+    <owners>$author$</owners>
+    <licenseUrl>https://github.com/nrjohnstone/AmbientContext/blob/master/LICENSE</licenseUrl> 
+    <projectUrl>https://github.com/nrjohnstone/AmbientContext</projectUrl>     
+    <requireLicenseAcceptance>false</requireLicenseAcceptance>
+    <description>$description$</description>
+    <releaseNotes></releaseNotes>
+    <copyright>Copyright 2016</copyright>
+    <tags></tags>
+  </metadata>
+
+</package>
